+++ .har/factory-line/cases/hapi-users/files/lib/infrastructure/webserver/server.js
@@ -1,0 +1,17 @@
+'use strict';
+
+const Hapi = require('@hapi/hapi');
+
+const createServer = async () => {
+  const server = Hapi.server({
+    port: process.env.PORT || 3000,
+  });
+
+  await server.register([
+    require('../../interfaces/routes/users'),
+  ]);
+
+  return server;
+};
+
+module.exports = createServer;